:recycle: Use BigDecimal for currency

+++ stadtbezirksbudget-backend/src/main/java/de/muenchen/stadtbezirksbudget/backend/antrag/entity/Finanzierung.java
@@ -7,6 +7,7 @@ import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 import java.io.Serial;
+import java.math.BigDecimal;
 import java.util.List;
 import lombok.Getter;
 import lombok.Setter;
@@ -27,7 +28,7 @@ public class Finanzierung extends BaseEntity {
     private boolean istWebsiteFoerderhinweis;
     private boolean istSonstigerFoerderhinweis;
     @NotNull private String sonstigeFoerderhinweise;
-    @PositiveOrZero private Double bewilligterZuschuss;
+    @PositiveOrZero private BigDecimal bewilligterZuschuss;
 
     @NotEmpty @OneToMany(mappedBy = "finanzierung")
     private List<VoraussichtlicheAusgabe> voraussichtlicheAusgaben;
